Small refactor of Revive of Revive and Death are communicated between classes via events and direct calls
Revive BP calls a single function on the playerstate.
PlayerState now has a reference to the full heal GE and applies it when told to revive.
PlayerState no longer directly calls functions on the controller when dying or reviving, controller instead listens to death and revive delegates on the character.