Card Level Caps, placeholder card.
Available cards that can be drawn are filtered depending on if the player has the max level of the card already.
Added a placeholder card to appear if the player draws no cards. (unlikely to happen once content is added, but useful for developing)
Removed old tags for upgrades.
Renamed Hearty upgrade to Toughness.